# openapi/specops-api.yaml
openapi: 3.1.0
info:
  title: SpecOps MVP API
  version: 1.0.0
  summary: MVP API contract for SpecOps parsing, review, traceability, locking, and export workflows.
  description: >
    This OpenAPI spec covers the minimum API surface required for the SpecOps MVP.
    It intentionally excludes advanced auto-fix orchestration, Signal-Sync automation,
    variant synchronization, and deep automotive workflow extensions.
servers:
  - url: https://api.specops.local
    description: Example local deployment
...
paths:
  /spec-sections:
    post:
      tags: [Spec Sections]
      summary: Upsert a parsed spec section
      operationId: upsertSpecSection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpecSection'
      responses:
        '201':
          description: Spec section created or updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpecSection'
        '400':
          $ref: '#/components/responses/BadRequest'
    get:
      tags: [Spec Sections]
      summary: List spec sections
      operationId: listSpecSections
      parameters:
        - $ref: '#/components/parameters/StatusParam'
        - in: query
          name: section_key
          schema:
            type: string
      responses:
        '200':
          description: Spec section list
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/SpecSection'
  /spec-sections/{specSectionId}:
    get:
      tags: [Spec Sections]
      summary: Get a spec section by id
      operationId: getSpecSection
      parameters:
        - $ref: '#/components/parameters/SpecSectionIdParam'
      responses:
        '200':
          description: Spec section
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpecSection'
        '404':
          $ref: '#/components/responses/NotFound'
  /notes:
    post:
      tags: [Notes]
      summary: Create a note from parsed spec sections
      operationId: createNote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNoteRequest'
      responses:
        '201':
          description: Note created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    get:
      tags: [Notes]
      summary: List notes
      operationId: listNotes
      parameters:
        - $ref: '#/components/parameters/StatusParam'
        - in: query
          name: source_spec_id
          schema:
            type: string
      responses:
        '200':
          description: Note list
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Note'
  /notes/{noteId}:
    get:
      tags: [Notes]
      summary: Get a note by id
      operationId: getNote
      parameters:
        - in: path
          name: noteId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Note
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '404':
          $ref: '#/components/responses/NotFound'
  /requirements:
    post:
      tags: [Requirements]
      summary: Create a requirement draft
      operationId: createRequirement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRequirementRequest'
      responses:
        '201':
          description: Requirement created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Requirement'
        '400':
          $ref: '#/components/responses/BadRequest'
    get:
      tags: [Requirements]
      summary: List requirements
      operationId: listRequirements
      parameters:
        - $ref: '#/components/parameters/StatusParam'
        - in: query
          name: source_spec_id
          schema:
            type: string
        - in: query
          name: review_queue
          schema:
            type: string
            enum: [all, impacted, manual_recovery, pending_review]
      responses:
        '200':
          description: Requirement list
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Requirement'
  /requirements/{requirementId}:
    get:
      tags: [Requirements]
      summary: Get a requirement by id
      operationId: getRequirement
      parameters:
        - $ref: '#/components/parameters/RequirementIdParam'
      responses:
        '200':
          description: Requirement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Requirement'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Requirements]
      summary: Update editable fields of a non-approved requirement
      operationId: patchRequirement
      parameters:
        - $ref: '#/components/parameters/RequirementIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchRequirementRequest'
      responses:
        '200':
          description: Requirement updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Requirement'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Requirement cannot be patched in current state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /test-requirements:
    post:
      tags: [Test Requirements]
      summary: Create a test requirement from requirements
      operationId: createTestRequirement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTestRequirementRequest'
      responses:
        '201':
          description: Test requirement created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestRequirement'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    get:
      tags: [Test Requirements]
      summary: List test requirements
      operationId: listTestRequirements
      parameters:
        - $ref: '#/components/parameters/StatusParam'
        - in: query
          name: source_requirement_id
          schema:
            type: string
      responses:
        '200':
          description: Test requirement list
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/TestRequirement'
  /test-requirements/{testRequirementId}:
    get:
      tags: [Test Requirements]
      summary: Get a test requirement by id
      operationId: getTestRequirement
      parameters:
        - in: path
          name: testRequirementId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Test requirement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestRequirement'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Test Requirements]
      summary: Update editable fields of a non-approved test requirement
      operationId: patchTestRequirement
      parameters:
        - in: path
          name: testRequirementId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchTestRequirementRequest'
      responses:
        '200':
          description: Test requirement updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestRequirement'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Test requirement cannot be patched in current state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /test-requirements/{testRequirementId}/submit-review:
    post:
      tags: [Test Requirements]
      summary: Submit a test requirement for review
      operationId: submitTestRequirementReview
      parameters:
        - in: path
          name: testRequirementId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitReviewRequest'
      responses:
        '200':
          description: Test requirement moved to IN_REVIEW
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestRequirement'
        '409':
          description: Invalid state transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /audit-rationales:
    post:
      tags: [Audit Rationales]
      summary: Create an audit rationale for a requirement
      operationId: createAuditRationale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAuditRationaleRequest'
      responses:
        '201':
          description: Audit rationale created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditRationale'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /audit-rationales/{auditRationaleId}:
    get:
      tags: [Audit Rationales]
      summary: Get an audit rationale by id
      operationId: getAuditRationale
      parameters:
        - in: path
          name: auditRationaleId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Audit rationale
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditRationale'
        '404':
          $ref: '#/components/responses/NotFound'
  /requirements/{requirementId}/submit-review:
    post:
      tags: [Requirements]
      summary: Submit a requirement for review
      operationId: submitRequirementReview
      parameters:
        - $ref: '#/components/parameters/RequirementIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitReviewRequest'
      responses:
        '200':
          description: Requirement moved to IN_REVIEW
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Requirement'
        '409':
          description: Invalid state transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /reviews:
    post:
      tags: [Reviews]
      summary: Record an approve or reject decision
      operationId: createReview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'
      responses:
        '201':
          description: Review recorded and artifact state updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewDecisionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Invalid state transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /reviews/{reviewId}:
    get:
      tags: [Reviews]
      summary: Get a review record
      operationId: getReview
      parameters:
        - $ref: '#/components/parameters/ReviewIdParam'
      responses:
        '200':
          description: Review record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewRecord'
        '404':
          $ref: '#/components/responses/NotFound'
  /locks/{sectionKey}:
    post:
      tags: [Locks]
      summary: Acquire or renew a section lock
      operationId: acquireOrRenewLock
      parameters:
        - $ref: '#/components/parameters/SectionKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LockRequest'
      responses:
        '200':
          description: Lock acquired or renewed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SectionLock'
        '409':
          description: Lock conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LockConflictResponse'
    delete:
      tags: [Locks]
      summary: Release a section lock
      operationId: releaseLock
      parameters:
        - $ref: '#/components/parameters/SectionKeyParam'
        - in: query
          name: owner_id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Lock released
        '404':
          $ref: '#/components/responses/NotFound'
  /trace/{artifactId}:
    get:
      tags: [Trace]
      summary: Get upstream and downstream traceability for an artifact
      operationId: getTraceMap
      parameters:
        - $ref: '#/components/parameters/ArtifactIdParam'
      responses:
        '200':
          description: Traceability graph entry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraceMapEntry'
        '404':
          $ref: '#/components/responses/NotFound'
  /exports/codebeamer:
    post:
      tags: [Exports]
      summary: Export an approved requirement to Codebeamer
      operationId: exportRequirementToCodebeamer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CodebeamerExportRequest'
      responses:
        '200':
          description: Export job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportJobResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Artifact is not exportable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  parameters:
    SpecSectionIdParam:
      in: path
      name: specSectionId
      required: true
      schema:
        type: string
    RequirementIdParam:
      in: path
      name: requirementId
      required: true
      schema:
        type: string
    ReviewIdParam:
      in: path
      name: reviewId
      required: true
      schema:
        type: string
    SectionKeyParam:
      in: path
      name: sectionKey
      required: true
      schema:
        type: string
...
    StatusParam:
      in: query
      name: status
      schema:
        $ref: '#/components/schemas/ArtifactStatus'
  responses:
    BadRequest:
      description: Invalid request payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    SpecSection:
      type: object
      additionalProperties: false
      required:
        - id
        - artifact_type
        - schema_version
        - status
        - version
        - section_key
        - title
        - text
        - source_refs
        - created_at
        - updated_at
      properties:
        id:
          type: string
        artifact_type:
          type: string
          const: spec_section
        schema_version:
          type: string
        status:
          type: string
          enum: [DRAFT, APPROVED, ARCHIVED]
        version:
          type: string
        section_key:
          type: string
        title:
          type: string
        text:
          type: string
        normalized_text:
          type:
            - string
            - 'null'
        parser_warnings:
          type: array
          items:
            type: string
        source_refs:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/SourceRef'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Note:
      type: object
      additionalProperties: false
      required:
        - id
        - artifact_type
        - schema_version
        - status
        - version
        - title
        - summary
        - source_spec_ids
        - created_at
        - updated_at
      properties:
        id:
          type: string
        artifact_type:
          type: string
          const: note
        schema_version:
          type: string
        status:
          $ref: '#/components/schemas/ArtifactStatus'
        version:
          type: string
        title:
          type: string
        summary:
          type: string
        source_spec_ids:
          type: array
          minItems: 1
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Requirement:
      type: object
      additionalProperties: false
      required:
        - id
        - artifact_type
        - schema_version
        - status
        - version
        - title
        - statement
        - source_spec_ids
        - source_note_ids
        - trace
        - created_at
        - updated_at
      properties:
        id:
          type: string
        artifact_type:
          type: string
          const: requirement
        schema_version:
          type: string
        status:
          $ref: '#/components/schemas/ArtifactStatus'
        version:
          type: string
        title:
          type: string
        statement:
          type: string
        source_spec_ids:
          type: array
          minItems: 1
          items:
            type: string
        source_note_ids:
          type: array
          minItems: 1
          items:
            type: string
        compliance:
          $ref: '#/components/schemas/Compliance'
        trace:
          type: object
          additionalProperties: false
          required: [graph_node_id, downstream_test_requirement_ids]
          properties:
            graph_node_id:
              type: string
            downstream_test_requirement_ids:
              type: array
              items:
                type: string
        audit_rationale_id:
          type:
            - string
            - 'null'
        variant_scope:
          type: string
          enum: [base, overlay]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateNoteRequest:
      type: object
      additionalProperties: false
      required: [title, summary, source_spec_ids]
      properties:
        title:
          type: string
        summary:
          type: string
        source_spec_ids:
          type: array
          minItems: 1
          items:
            type: string
    CreateRequirementRequest:
      type: object
      additionalProperties: false
      required: [title, statement, source_spec_ids, source_note_ids, variant_scope]
      properties:
        title:
          type: string
        statement:
          type: string
        source_spec_ids:
          type: array
          minItems: 1
          items:
            type: string
        source_note_ids:
          type: array
          minItems: 1
          items:
            type: string
        compliance:
          $ref: '#/components/schemas/Compliance'
        variant_scope:
          type: string
          enum: [base, overlay]
    PatchRequirementRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        title:
          type: string
        statement:
          type: string
        compliance:
          $ref: '#/components/schemas/Compliance'
        audit_rationale_id:
          type:
            - string
            - 'null'
    TestRequirement:
      type: object
      additionalProperties: false
      required:
        - id
        - artifact_type
        - schema_version
        - status
        - version
        - statement
        - source_requirement_ids
        - acceptance_criteria
        - created_at
        - updated_at
      properties:
        id:
          type: string
        artifact_type:
          type: string
          const: test_requirement
        schema_version:
          type: string
        status:
          $ref: '#/components/schemas/ArtifactStatus'
        version:
          type: string
        statement:
          type: string
        source_requirement_ids:
          type: array
          minItems: 1
          items:
            type: string
        acceptance_criteria:
          type: array
          items:
            type: string
        audit_rationale_id:
          type:
            - string
            - 'null'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateTestRequirementRequest:
      type: object
      additionalProperties: false
      required: [statement, source_requirement_ids]
      properties:
        statement:
          type: string
        source_requirement_ids:
          type: array
          minItems: 1
          items:
            type: string
        acceptance_criteria:
          type: array
          items:
            type: string
        audit_rationale_id:
          type:
            - string
            - 'null'
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
        detail:
          type:
            - string
            - 'null'